Add a basic compiler
It can currently only handle opperations with integer.
It only supports functions declared at the global scope and does not
support callbacks or returning functions from other functions.
It should support the same operators that the interpreted version
supports although only with integers.
Currently the main function of the compiled code is called when
running the compiler, next step is to write the compiled version to an
executable file.