Refactor and add an environment to every expression
The environment tracks type information such as the types the
expression is executed with to allow for future type narrowing and
tracks the resulting type of the expression.
This allows for far better efficiency and cleaner code, not having to
call a function that returns a result for an already known value for example.