Remove syntactic requirement for typeof expressions to need to be wrapped in brackets when parts of some other expressions
Note that for them to be useful they still might need to be wrapped in brackets as they are greedy and will attempt to get the type of the complete expression directly following them.
For example the brackets are needed in
if (typeof x)==typeof null {}else{}
Otherwise it would be evaluated as
if typeof (x==typeof null) {}else{}
which is a type error as the result is not a Bool.